+++ milkman/src/main/java/milkman/exporter/CustomTemplateExporter.java
@@ -1,23 +1,26 @@
 package milkman.exporter;
 
-import static milkman.utils.fxml.FxmlBuilder.choiceBox;
-import static milkman.utils.fxml.FxmlBuilder.vbox;
+import static milkman.utils.fxml.FxmlBuilder.comboBox;
+import static milkman.utils.fxml.FxmlBuilder.hbox;
+import static milkman.utils.fxml.FxmlBuilder.label;
+import static milkman.utils.fxml.FxmlBuilder.space;
 
+import com.jfoenix.controls.JFXComboBox;
 import java.util.List;
 import javafx.scene.Node;
-import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.TextArea;
-import javafx.scene.layout.HBox;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
 import milkman.domain.RequestContainer;
 import milkman.exporter.CustomTemplateService.CompiledTemplate;
 import milkman.ui.main.Toaster;
 import milkman.ui.plugin.RequestExporterPlugin;
+import milkman.ui.plugin.RequestTypePlugin;
+import milkman.ui.plugin.RequestTypePluginAware;
 import milkman.ui.plugin.Templater;
 
 
-public class CustomTemplateExporter implements RequestExporterPlugin {
+public class CustomTemplateExporter implements RequestExporterPlugin, RequestTypePluginAware {
 
 
   private CustomTemplateService templateService = new CustomTemplateService();
@@ -25,7 +28,8 @@ public class CustomTemplateExporter implements RequestExporterPlugin {
 
   private RequestContainer request;
   private List<CompiledTemplate> exporters;
-  private ChoiceBox<CompiledTemplate> cbTemplate;
+  private JFXComboBox<CompiledTemplate> cbTemplate;
+  private List<RequestTypePlugin> registeredRequestTypePlugins;
 
   @Override
   public String getName() {
@@ -35,20 +39,23 @@ public class CustomTemplateExporter implements RequestExporterPlugin {
   @Override
   public Node getRoot(RequestContainer request, Templater templater) {
     this.request = request;
-    this.exporters = templateService.getTemplatesForType(request.getType());
+    String requestType = registeredRequestTypePlugins.stream().filter(r -> r.canHandle(request))
+        .map(r -> r.getRequestType()).findAny().orElse("");
+    this.exporters = templateService.getTemplatesForType(requestType);
     textArea = new TextArea();
     textArea.setEditable(false);
 
-    cbTemplate = choiceBox("templateName");
+    cbTemplate = comboBox("templateName");
     cbTemplate.getItems().addAll(exporters);
     cbTemplate.valueProperty().addListener((obs, o, n) -> {
       if (n != null) {
         refreshCommand(templater, n);
       }
     });
+    cbTemplate.getSelectionModel().selectFirst();
 
     VBox root = new VBox();
-    root.getChildren().add(cbTemplate);
+    root.getChildren().add(hbox(label("Template: "), space(5), cbTemplate));
     root.getChildren().add(textArea);
     VBox.setVgrow(textArea, Priority.ALWAYS);
     return 	root;
@@ -77,5 +84,10 @@ public class CustomTemplateExporter implements RequestExporterPlugin {
   @Override
   public boolean canHandle(RequestContainer request) {
     return templateService.canHandleRequestType(request.getType());
+  }
+
+  @Override
+  public void setRequestTypePlugins(List<RequestTypePlugin> plugins) {
+    this.registeredRequestTypePlugins = plugins;
   }
 }