+++ milkman/src/main/java/milkman/ui/plugin/RequestTypePluginAware.java
@@ -1,0 +1,14 @@
+package milkman.ui.plugin;
+
+import java.util.List;
+import milkman.domain.KeySet;
+
+/**
+ * 
+ * @author pmucha
+ *
+ */
+public interface RequestTypePluginAware {
+
+	void setRequestTypePlugins(List<RequestTypePlugin> plugins);
+}