+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
-versuib 5.6.0 -- upcomming
-  * 
+version 5.6.0 -- upcomming (nightly)
+  * custom export templates 
 
 version 5.5.0